test(waterfall): rename sort control test to avoid path collision
Master added test/Waterfall/controlPanel.test.ts in #42533; rename the
x-axis-sort control test to controlPanel.sort.test.ts so the two files
no longer collide (add/add conflict).
